sprite handles: type SpriteIcon/TextSprite fields, local-cast numeric sites
- icon.h: SpriteIcon.field_18 holds a (dark) sprite, only used by title.c;
  type it struct Sprite* and clear with NULL.
- text.c: TextSprite.field_1c is a sprite passed to PrintSprite; type it
  struct Sprite*.
- jungle_cruise.c: DAT_0082c6a0 is a shared int handle (returned by &, also
  set from non-sprite sources) -> keep int, cast the sprite source.
- challenge.c FUN_00445000: raw offset-walked sprite handles read as
  unsigned int -> cast at the KillSprite boundary.

Byte-identical: InitTitleScreen/FUN_00434740/FUN_00455ec0/FUN_00445000 100%,
aggregate 53.02%.

+++ src/legoland/jungle_cruise.c
@@ -238,7 +238,7 @@ unsigned int *FUN_00434740(struct JungleFishHolder *param_1, unsigned short para
         lls2 = (struct JungleLLS *)GetLLSForSprite((struct SpriteLLS *)DAT_0081cb6c);
         frame = lls1->field_0;
         LLSSetFrame((struct LLS *)lls2, frame); /* TODO: fold struct JungleLLS into struct LLS */
-        DAT_0082c6a0 = DAT_0081cb6c;
+        DAT_0082c6a0 = (unsigned int)DAT_0081cb6c;
     } else {
         DAT_0082c6a0 = obj->field_64;
     }
